# docs/swagger/api/v1/greet.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/greet.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "GreeterService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/greet/chatty": {
      "get": {
        "summary": "Client Streaming",
        "operationId": "GreeterService_ChattyClientSaysHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DelayedReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GreeterService"
        ]
      }
    },
    "/v1/greet/hello": {
      "get": {
        "summary": "Unary",
        "operationId": "GreeterService_SayHello2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GreeterService"
        ]
      },
      "post": {
        "summary": "Unary",
        "operationId": "GreeterService_SayHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "The request message containing the user's name.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1HelloRequest"
            }
          }
        ],
        "tags": [
          "GreeterService"
        ]
      }
    },
    "/v1/greet/interacting": {
      "get": {
        "summary": "Both Streaming",
        "operationId": "GreeterService_InteractingHello",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1HelloResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1HelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GreeterService"
        ]
      }
    },
    "/v1/greet/parrot": {
      "get": {
        "summary": "Server Streaming",
        "operationId": "GreeterService_ParrotSaysHello",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1HelloResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1HelloResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "greeting",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GreeterService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1DelayedReply": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "request": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1HelloRequest"
          }
        }
      }
    },
    "v1HelloRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "greeting": {
          "type": "string"
        }
      },
      "description": "The request message containing the user's name."
    },
    "v1HelloResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "description": "The response message containing the greetings."
    }
  }
}